requirements.txt
streamlit 
numpy
matplotli